requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2
python-dateutil==2.9.0.post0
feedparser==6.0.11
